+++ mini_template_macro/tests/create_modifier/unused_arg.rs
@@ -1,0 +1,8 @@
+use mini_template_macro::create_modifier;
+
+#[create_modifier]
+fn modifier(value: String, _: usize) -> String {
+    value
+}
+
+fn main() {}